Update watermark_app.py
Key Features Added:
Watermark Text Position:

Top: Places the watermark at the top.

Middle: Places the watermark in the middle.

Bottom: Places the watermark at the bottom.

The position is set dynamically based on the user's selection in the st.selectbox.

Watermark Opacity:

The opacity of the watermark text can be adjusted using a slider (0.0 to 1.0), where 0.0 is completely transparent and 1.0 is fully opaque.

Watermark Color:

The color of the watermark text can be selected using a color picker. You can pick any color, and the script will convert it into the appropriate RGB format for use.